feat:tutorial-4-21cineplux (#9)
* feat:tutorial-4-21cineplux

* solusi tut4 apap

+++ 21cineplux/src/main/java/apap/tutorial/cineplux/service/PenjagaServiceImpl.java
@@ -1,11 +1,15 @@
 package apap.tutorial.cineplux.service;
 
+import apap.tutorial.cineplux.model.BioskopModel;
 import apap.tutorial.cineplux.model.PenjagaModel;
+import apap.tutorial.cineplux.repository.BioskopDB;
 import apap.tutorial.cineplux.repository.PenjagaDB;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.time.LocalTime;
+import java.util.Optional;
 
 @Service
 @Transactional
@@ -13,18 +17,33 @@ public class PenjagaServiceImpl implements PenjagaService {
 
     @Autowired
     PenjagaDB penjagaDB;
-    
+
+    @Autowired
+    BioskopDB bioskopDB;
+
     @Override
     public void addPenjaga(PenjagaModel penjaga) {
         penjagaDB.save(penjaga);
     }
 
     @Override
-    public void updatePenjaga(PenjagaModel penjaga) { penjagaDB.save(penjaga); }
+    public PenjagaModel getPenjagaByNoPenjaga(Long noPenjaga) {
+        Optional<PenjagaModel> penjaga = penjagaDB.findByNoPenjaga(noPenjaga);
+        return penjaga.orElse(null);
+    }
 
     @Override
-    public void deletePenjaga(PenjagaModel penjaga) {
-        penjagaDB.delete(penjaga);
-        System.out.println("masuk ke detelte");
+    public void updatePenjaga(PenjagaModel penjaga) {
+        penjagaDB.save(penjaga);
+    }
+
+    @Override public int deletePenjaga(PenjagaModel penjaga) {
+        LocalTime now = LocalTime.now();
+        BioskopModel bioskop = bioskopDB.findByNoBioskop(penjaga.getBioskop().getNoBioskop()).get();
+        if (now.isBefore(bioskop.getWaktuBuka()) || now.isAfter(bioskop.getWaktuTutup())) {
+            penjagaDB.delete(penjaga);
+            return 1;
+        }
+        return 0;
     }
 }
